moved search function to dino-search.js
